requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
langgraph
faiss-cpu
streamlit